requirements.txt
minimalmodbus==2.1.1 
pyserial==3.5
matplotlib==3.9.2
pymodbus==3.6.5
ttkbootstrap==1.10.1
